Fix max on duration

+++ src/BedJoint.h
@@ -13,7 +13,7 @@ class BedJoint {
 	private:
 		int _pinUp;
 		int _pinDown;
-		uint32_t _max;          // Maximum milliseconds to remain on
+		uint64_t _max;          // Maximum milliseconds to remain on
 		uint64_t _started;      // Milliseconds when the movement started
 		uint64_t _target;       // Milliseconds when the movement is to stop
 		uint64_t _rampUp;       // Milliseconds to ramp up to full speed
